requirements.txt
djangorestframework-3.14.0 pytz-2022.7.1
django-environ-0.9.0
asgiref-3.6.0
django-4.1.6
sqlparse-0.4.3